+++ pkg/controller/core.oam.dev/v1alpha2/application/suite_test.go
@@ -17,35 +17,21 @@ limitations under the License.
 package application
 
 import (
-	"context"
-	"encoding/json"
 	"path/filepath"
 	"testing"
-	"time"
-
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
-
-	"github.com/oam-dev/kubevela/pkg/oam/discoverymapper"
-	"github.com/oam-dev/kubevela/pkg/webhook/core.oam.dev/v1alpha2/applicationconfiguration"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/yaml"
 
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1alpha2"
-	"github.com/oam-dev/kubevela/pkg/oam/util"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -98,202 +84,3 @@ var _ = AfterSuite(func() {
 	err := testEnv.Stop()
 	Expect(err).ToNot(HaveOccurred())
 })
-
-var _ = Describe("Test Application Controller", func() {
-	ctx := context.Background()
-
-	ns := corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "vela-test",
-		},
-	}
-
-	app := &v1alpha2.Application{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Application",
-			APIVersion: "core.oam.dev/v1alpha2",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "application-sample",
-			Namespace: ns.Name,
-		},
-		Spec: v1alpha2.ApplicationSpec{
-			Components: []v1alpha2.ApplicationComponent{
-				{
-					Name:         "myweb",
-					WorkloadType: "worker",
-					Settings:     runtime.RawExtension{Raw: []byte("{\"cmd\":[\"sleep\",\"1000\"],\"image\":\"busybox\"}")},
-				},
-			},
-		},
-	}
-
-	wd := &v1alpha2.WorkloadDefinition{}
-	wDDefJson, _ := yaml.YAMLToJSON([]byte(wDDefYaml))
-
-	td := &v1alpha2.TraitDefinition{}
-	tDDefJson, _ := yaml.YAMLToJSON([]byte(tDDefYaml))
-
-	BeforeEach(func() {
-		Expect(json.Unmarshal(wDDefJson, wd)).Should(BeNil())
-		Expect(k8sClient.Create(ctx, wd)).Should(SatisfyAny(BeNil(), &util.AlreadyExistMatcher{}))
-
-		Expect(json.Unmarshal(tDDefJson, td)).Should(BeNil())
-		Expect(k8sClient.Create(ctx, td)).Should(SatisfyAny(BeNil(), &util.AlreadyExistMatcher{}))
-
-		Eventually(
-			func() error {
-				return k8sClient.Create(ctx, &ns)
-			},
-			time.Second*3, time.Millisecond*300).Should(SatisfyAny(BeNil(), &util.AlreadyExistMatcher{}))
-	})
-
-	It("Test Application Reconciler", func() {
-		Expect(k8sClient.Create(ctx, app.DeepCopyObject())).Should(BeNil())
-		reconciler := &Reconciler{
-			Client: k8sClient,
-			Log:    ctrl.Log.WithName("Application"),
-			Scheme: testScheme,
-		}
-		appKey := client.ObjectKey{
-			Name:      app.Name,
-			Namespace: app.Namespace,
-		}
-		reconcileRetry(reconciler, reconcile.Request{
-			NamespacedName: appKey})
-
-		checkApp := &v1alpha2.Application{}
-		Expect(k8sClient.Get(ctx, appKey, checkApp)).Should(BeNil())
-		Expect(checkApp.Status.Phase).Should(Equal(v1alpha2.ApplicationRunning))
-
-		appConfig := &v1alpha2.ApplicationConfiguration{}
-		Expect(k8sClient.Get(ctx, client.ObjectKey{
-			Namespace: app.Namespace,
-			Name:      app.Name,
-		}, appConfig)).Should(BeNil())
-
-		component := &v1alpha2.Component{}
-		Expect(k8sClient.Get(ctx, client.ObjectKey{
-			Namespace: app.Namespace,
-			Name:      "myweb",
-		}, component)).Should(BeNil())
-
-		dm, err := discoverymapper.New(cfg)
-		Expect(err).Should(BeNil())
-		decoder, err := admission.NewDecoder(testScheme)
-		Expect(err).Should(BeNil())
-		handler := applicationconfiguration.ValidatingHandler{
-			Client:  k8sClient,
-			Mapper:  dm,
-			Decoder: decoder,
-			Validators: []applicationconfiguration.AppConfigValidator{
-				applicationconfiguration.AppConfigValidateFunc(applicationconfiguration.ValidateTraitObjectFn),
-				applicationconfiguration.AppConfigValidateFunc(applicationconfiguration.ValidateRevisionNameFn),
-				applicationconfiguration.AppConfigValidateFunc(applicationconfiguration.ValidateWorkloadNameForVersioningFn),
-				applicationconfiguration.AppConfigValidateFunc(applicationconfiguration.ValidateTraitAppliableToWorkloadFn),
-				applicationconfiguration.AppConfigValidateFunc(applicationconfiguration.ValidateTraitConflictFn),
-				// TODO(wonderflow): Add more validation logic here.
-			},
-		}
-
-		vAppConfig := &applicationconfiguration.ValidatingAppConfig{}
-		if err := vAppConfig.PrepareForValidation(ctx, k8sClient, dm, appConfig); err != nil {
-			Expect(err).Should(BeNil())
-		}
-		for _, validator := range handler.Validators {
-			allErrs := validator.Validate(ctx, *vAppConfig)
-			Expect(len(allErrs)).Should(BeEquivalentTo(0))
-		}
-
-	})
-
-})
-
-func reconcileRetry(r reconcile.Reconciler, req reconcile.Request) {
-	Eventually(func() error {
-		_, err := r.Reconcile(req)
-		return err
-	}, 3*time.Second, time.Second).Should(BeNil())
-}
-
-const (
-	wDDefYaml = `
-apiVersion: core.oam.dev/v1alpha2
-kind: WorkloadDefinition
-metadata:
-  name: worker
-  annotations:
-    definition.oam.dev/description: "Long-running scalable backend worker without network endpoint"
-spec:
-  definitionRef:
-    name: deployments.apps
-  extension:
-    template: |
-      output: {
-      	apiVersion: "apps/v1"
-      	kind:       "Deployment"
-      	spec: {
-      		selector: matchLabels: {
-      			"app.oam.dev/component": context.name
-      		}
-
-      		template: {
-      			metadata: labels: {
-      				"app.oam.dev/component": context.name
-      			}
-
-      			spec: {
-      				containers: [{
-      					name:  context.name
-      					image: parameter.image
-
-      					if parameter["cmd"] != _|_ {
-      						command: parameter.cmd
-      					}
-      				}]
-      			}
-      		}
-
-      		selector:
-      			matchLabels:
-      				"app.oam.dev/component": context.name
-      	}
-      }
-
-      parameter: {
-      	// +usage=Which image would you like to use for your service
-      	// +short=i
-      	image: string
-
-      	cmd?: [...string]
-      }`
-	tDDefYaml = `
-apiVersion: core.oam.dev/v1alpha2
-kind: TraitDefinition
-metadata:
-  annotations:
-    definition.oam.dev/description: "Manually scale the app"
-  name: scaler
-spec:
-  appliesToWorkloads:
-    - webservice
-    - worker
-  definitionRef:
-    name: manualscalertraits.core.oam.dev
-  workloadRefPath: spec.workloadRef
-  extension:
-    template: |-
-      output: {
-      	apiVersion: "core.oam.dev/v1alpha2"
-      	kind:       "ManualScalerTrait"
-      	spec: {
-      		replicaCount: parameter.replicas
-      	}
-      }
-      parameter: {
-      	//+short=r
-      	replicas: *1 | int
-      }
-
-`
-)
